Refactor and use jump tables to compare cpu usage on session creation
