+++ onnxruntime/core/providers/qnn/builder/qnn_node_group.h
