+++ onnxruntime/core/providers/qnn/qnn_execution_provider.h
@@ -11,6 +11,7 @@
 #include "core/providers/qnn/builder/qnn_backend_manager.h"
 #include "core/providers/qnn/builder/qnn_model.h"
 #include "core/providers/qnn/builder/qnn_configs_helper.h"
+#include "core/providers/qnn/builder/qnn_node_group.h"
 #include "HTP/QnnHtpGraph.h"
 #include <vector>
 #include <set>
